expression: move some funcs out of builtin_other.
